// src/types/mangadex.ts
export type Relationship = {
  /** UUID formatted string */
  id: string;
  type: string;
  /** Only present if you are on a Manga entity and a Manga relationship */
  related:
    | "monochrome"
    | "main_story"
    | "adapted_from"
    | "based_on"
    | "prequel"
    | "side_story"
    | "doujinshi"
    | "same_franchise"
    | "shared_universe"
    | "sequel"
    | "spin_off"
    | "alternate_story"
    | "alternate_version"
    | "preserialization"
    | "colored"
    | "serialization";
  /** If Reference Expansion is applied, contains objects attributes */
  attributes: any | null;
};

export type Chapter = {
  /** UUID formatted string */
  id: string;
  type: "chapter";
  attributes: ChapterAttributes;
  relationships: Relationship[];
};

export type Manga = {
  /** UUID formatted string */
  id: string;
  type: "manga";
  attributes: MangaAttributes;
  relationships: Relationship[];
};
// ...
export type MangaAttributes = {
  title: LocalizedString;
  altTitles: LocalizedString[];
  description: LocalizedString;
  isLocked: boolean;
  links: Links | null;
  originalLanguage: string;
  lastVolume: string | null;
  lastChapter: string | null;
  publicationDemographic: "shounen" | "shoujo" | "josei" | "seinen" | null;
  status: "completed" | "ongoing" | "cancelled" | "hiatus";
  year: number | null;
  contentRating: "safe" | "suggestive" | "erotica" | "pornographic";
  chapterNumbersResetOnNewVolume: boolean;
  availableTranslatedLanguages: any[];
  /** UUID formatted string */
  latestUploadedChapter: string;
  tags: Tag[];
  state: "draft" | "submitted" | "published" | "rejected";
  /** ```console
   * Minimum: 1
   * ``` */
  version: number;
  createdAt: string;
  updatedAt: string;
};
// ...
export type Cover = {
  /** UUID formatted string */
  id: string;
  type: "cover_art";
  attributes: CoverAttributes;
  relationships: Relationship[];
};

export type CoverAttributes = {
  volume: string | null;
  fileName: string;
  description: string | null;
  locale: string | null;
  /** ```console
   * Minimum: 1
   * ``` */
  version: number;
  createdAt: string;
  updatedAt: string;
};
// ...
export type Author = {
  /** UUID formatted string */
  id: string;
  type: "author";
  attributes: AuthorAttributes;
  relationships: Relationship[];
};

export type AuthorAttributes = {
  name: string;
  imageUrl: string;
  biography: LocalizedString;
  /** Pattern: ^https?:\/\/twitter\.com(\/|$) */
  twitter: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?pixiv\.net(\/|$) */
  pixiv: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?melonbooks\.co\.jp(\/|$) */
  melonBook: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?fanbox\.cc(\/|$) */
  fanBox: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?booth\.pm(\/|$) */
  booth: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?nicovideo\.jp(\/|$) */
  nicoVideo: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?skeb\.jp(\/|$) */
  skeb: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?fantia\.jp(\/|$) */
  fantia: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?tumblr\.com(\/|$) */
  tumblr: string | null;
  /** Pattern: ^https?:\/\/www\.youtube\.com(\/|$) */
  youtube: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?weibo\.(cn|com)(\/|$) */
  weibo: string | null;
  /** Pattern: ^https?:\/\/([\w-]+\.)?naver\.com(\/|$) */
  naver: string | null;
  /** Pattern: ^https?:\/\/ */
  website: string | null;
  /** ```console
   * Minimum: 1
   * ``` */
  version: number;
  createdAt: string;
  updatedAt: string;
};
// ...
export type ExtendManga = Manga & {
  cover_art?: Partial<Cover> & Pick<Cover, "id" | "type">;
  author?: Partial<Author> & Pick<Author, "id" | "type">;
  artist?: Partial<Author> & Pick<Author, "id" | "type">;
};